Moving the query creator to specific class
