+++ src/ElasticSearch.php
@@ -41,6 +41,9 @@ class ElasticSearch extends ConfigurableService implements Search
     /** @var \Elasticsearch\Client */
     private $client;
 
+    /** @var QueryBuilder */
+    private $queryBuilder;
+
     /** @return \Elasticsearch\Client */
     protected function getClient(): Client
     {
@@ -51,6 +54,16 @@ class ElasticSearch extends ConfigurableService implements Search
         }
 
         return $this->client;
+    }
+
+    /** @return QueryBuilder */
+    protected function getQueryBuilder(): QueryBuilder
+    {
+        if (is_null($this->queryBuilder)) {
+            $this->queryBuilder = QueryBuilder::create();
+        }
+
+        return $this->queryBuilder;
     }
 
     protected function getIndexer(): ElasticSearchIndexer
@@ -77,7 +90,7 @@ class ElasticSearch extends ConfigurableService implements Search
         try {
             return $this->buildResultSet(
                 $this->getClient()->search(
-                    $this->getSearchParams($queryString, $type, $start, $count, $order, $dir)
+                    $this->getQueryBuilder()->getSearchParams($queryString, $type, $start, $count, $order, $dir)
                 )
             );
         } catch (\Exception $e) {
@@ -140,56 +153,6 @@ class ElasticSearch extends ConfigurableService implements Search
     }
 
     /**
-     * @param string $queryString
-     * @param string $type
-     * @param int $start
-     * @param int $count
-     * @param $order
-     * @param $dir
-     * @return array
-     */
-    protected function getSearchParams($queryString, $type, $start, $count, $order, $dir)
-    {
-        $parts = explode(' ', htmlspecialchars_decode($queryString));
-
-        foreach ($parts as $key => $part) {
-            $matches = [];
-            $part = $this->updateIfUri($part);
-            if (preg_match('/^([^a-z_]*)([a-z_]+):(.*)/', $part, $matches) === 1) {
-                [$fullstring, $prefix, $fieldname, $value] = $matches;
-                $value = $this->updateIfUri($value);
-                if ($fieldname) {
-                    $parts[$key] = $prefix . $fieldname . ':' . str_replace(':', '\\:', $value);
-                }
-            }
-        }
-        $queryString = implode(' ', $parts);
-        $queryString = (strlen($queryString) == 0 ? '' : '(' . $queryString . ') AND ')
-            . 'type:' . str_replace(':', '\\:', '"' . $type . '"');
-
-        $query = [
-            'query' => [
-                'query_string' =>
-                    [
-                        "default_operator" => "AND",
-                        "query" => $queryString
-                    ]
-            ],
-            'sort' => [$order => ['order' => $dir]]
-        ];
-
-        $params = [
-            "index" => implode(',', IndexerInterface::AVAILABLE_INDEXES), //TODO we need to specificy only one index during implementation of task (TAO-10248)
-            "size" => $count,
-            "from" => $start,
-            "client" => ["ignore" => 404],
-            "body" => json_encode($query)
-        ];
-
-        return $params;
-    }
-
-    /**
      * @param array $elasticResult
      * @return ResultSet
      */
@@ -209,13 +172,5 @@ class ElasticSearch extends ConfigurableService implements Search
         $uris = array_unique($uris);
 
         return new ResultSet($uris, $total);
-    }
-
-    protected function updateIfUri(string $query): string
-    {
-        if (\common_Utils::isUri($query)) {
-            $query = '"' . $query . '"';
-        }
-        return $query;
     }
 }
